+++ src/main/java/chikachi/discord/ChikachiDiscord.java
@@ -1,8 +1,6 @@
 package chikachi.discord;
 
-import chikachi.discord.command.ChikachiDiscordCommand;
-import net.minecraft.command.ICommandSender;
-import net.minecraft.util.ChatComponentText;
+import chikachi.discord.command.NonLibCommandHandler;
 import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.*;
@@ -14,34 +12,30 @@ import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
-@SuppressWarnings("WeakerAccess")
 @Mod(modid = Constants.MODID, name = Constants.MODNAME, version = Constants.VERSION, serverSideOnly = true, acceptableRemoteVersions = "*")
 public class ChikachiDiscord {
     private static Proxy proxy = new Proxy();
 
     private static final Logger logger = LogManager.getLogger(Constants.MODID);
 
-    @SuppressWarnings("unused")
     @Mod.EventHandler
     public void onPreInit(FMLPreInitializationEvent event) {
         proxy.onPreInit(event);
     }
 
-    @SuppressWarnings("unused")
     @Mod.EventHandler
     public void onServerAboutToStart(FMLServerAboutToStartEvent event) {
         proxy.onServerAboutToStart(event);
     }
 
-    @SuppressWarnings("unused")
     @Mod.EventHandler
     public void onServerStarting(FMLServerStartingEvent event) {
         proxy.onServerStarting(event);
 
         if (Loader.isModLoaded("ChikachiLib")) {
-            Log("Trying to hook into ChikachiLib");
+            Log("Trying to hook into ChikachiLib", false);
             try {
-                Class subCommandHandlerClass = Class.forName("chikachi.discord.command.ChikachiLibIntegration");
+                Class subCommandHandlerClass = Class.forName("chikachi.discord.command.LibCommandHandler");
                 Class libClass = Class.forName("chikachi.lib.ChikachiLib");
                 Class commandClass = Class.forName("chikachi.lib.common.command.sub.CommandChikachiBase");
                 if (subCommandHandlerClass != null && libClass != null && commandClass != null) {
@@ -52,61 +46,36 @@ public class ChikachiDiscord {
                             Method registerCommandHandlerMethod = obj.getClass().getMethod("RegisterSubCommandHandler", commandClass);
                             if (registerCommandHandlerMethod != null) {
                                 registerCommandHandlerMethod.invoke(obj, subCommandHandlerClass.newInstance());
-                                Log("Hooked into ChikachiLib");
+                                Log("Hooked into ChikachiLib", false);
                                 return;
                             }
                         }
                     }
                 }
             } catch (ClassNotFoundException | NoSuchFieldException | NoSuchMethodException | IllegalAccessException | InstantiationException | InvocationTargetException e) {
-                Log("Failed to hook into ChikachiLib");
+                Log("Failed to hook into ChikachiLib", false);
             }
         }
 
-        event.registerServerCommand(new ChikachiDiscordCommand());
+        event.registerServerCommand(new NonLibCommandHandler());
     }
 
-    @SuppressWarnings("unused")
     @Mod.EventHandler
     public void onServerStarted(FMLServerStartedEvent event) {
         proxy.onServerStarted(event);
     }
 
-    @SuppressWarnings("unused")
     @Mod.EventHandler
     public void onServerShutdown(FMLServerStoppingEvent event) {
         proxy.onServerShutdown(event);
     }
 
-    public static void processCommand(ICommandSender sender, String[] args) {
-        if (args.length > 0) {
-            switch (args[0]) {
-                case "reload":
-                    Configuration.load();
-                    DiscordClient.getInstance().disconnect();
-                    DiscordClient.getInstance().connect();
-                    sender.addChatMessage(new ChatComponentText("Config reloaded"));
-                    return;
-                case "start":
-                    DiscordClient.getInstance().connect();
-                    sender.addChatMessage(new ChatComponentText("Connected"));
-                    return;
-                case "stop":
-                    DiscordClient.getInstance().disconnect();
-                    sender.addChatMessage(new ChatComponentText("Disconnected"));
-                    return;
-            }
-        }
-
-        sender.addChatMessage(new ChatComponentText("Unknown command - Available commands: reload, start, stop"));
-    }
-
-    @SuppressWarnings("unused")
+    @SuppressWarnings("WeakerAccess")
     public static void Log(String message) {
         Log(message, false);
     }
 
-    static void Log(String message, boolean warning) {
+    public static void Log(String message, boolean warning) {
         logger.log(warning ? Level.WARN : Level.INFO, String.format("[%s] %s", Constants.VERSION, message));
     }
 }