+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### 0.4.0
+- Added max length on relayed Discord messages (Default -1, meaning disabled)
+- Added limit of commands to roles (by role name)
+- Improved Discord user mentions (Prevents wrong mention in case of both `User1` and `User1_`)
+- Replaced placeholders in strings (**BREAKS ALL CURRENT CONFIGS**)
+	- `%USER%` Name of user (Chat, Death, Achievement, Join, Leave)
+	- `%MESSAGE%` Message (Chat, Death)
+	- `%ACHIEVEMENT%` Name of achievement (Achievement)
+- Returned player names from `!online` is now packed in like `this`
+- Remove Minecraft formatting (1 character prefixed with §) from `/say` and chat messages
+
 ### 0.3.0
 - Added `!tps` command for Discord
 - Highlight player's name in Discord messages
@@ -7,9 +18,9 @@
 
 ### 0.2.0
 - Added command
-	- If ChikachiLib is detected, /chikachi discord
-	- Otherwise, /discord
-- Broadcast messages (/say) is now also relayed
+	- If ChikachiLib is detected, `/chikachi discord`
+	- Otherwise, `/discord`
+- Broadcast messages (`/say`) is now also relayed
 - Made Discord mentions case-insensitive
 
 #### 0.1.3